+++ src/main/java/com/getpcpanel/commands/command/CommandVolumeProcessMute.java
@@ -1,6 +1,7 @@
 package com.getpcpanel.commands.command;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
+import java.util.Set;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.getpcpanel.cpp.MuteType;
 import com.getpcpanel.cpp.SndCtrl;
@@ -11,17 +12,16 @@ import lombok.ToString;
 @Getter
 @ToString(callSuper = true)
 public class CommandVolumeProcessMute extends CommandVolume implements ButtonAction {
-    private final String processName;
+    private final Set<String> processName;
     private final MuteType muteType;
 
-    @JsonCreator
-    public CommandVolumeProcessMute(@JsonProperty("processName") String processName, @JsonProperty("muteType") MuteType muteType) {
+    public CommandVolumeProcessMute(@JsonProperty("processName") Set<String> processName, @JsonProperty("muteType") MuteType muteType) {
         this.processName = processName;
         this.muteType = muteType;
     }
 
     @Override
     public void execute() {
-        SndCtrl.muteProcess(processName, muteType);
+        SndCtrl.muteProcesses(processName, muteType);
     }
 }