+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - You can set a main profile which will be loaded when you start the application
 - Added version checker. If a new version is released the UI will show that with a link to the download page.
 - More than 2 processes can be selected for volume control
+- More than 1 process can be selected for muting
 
 **Fixes:**